+++ glyph/src/font_arc.rs
@@ -140,6 +140,16 @@ impl Font for FontArc {
     fn glyph_raster_image2(&self, id: GlyphId, size: u16) -> Option<v2::GlyphImage> {
         self.0.glyph_raster_image2(id, size)
     }
+
+    #[inline]
+    fn glyph_svg_image(&self, id: GlyphId) -> Option<crate::GlyphSvg> {
+        self.0.glyph_svg_image(id)
+    }
+
+    #[inline]
+    fn font_data(&self) -> &[u8] {
+        self.0.font_data()
+    }
 }
 
 impl From<FontVec> for FontArc {