+++ glyph/src/ttfp.rs
@@ -3,7 +3,7 @@ mod outliner;
 #[cfg(feature = "variable-fonts")]
 mod variable;
 
-use crate::{point, v2, Font, GlyphId, GlyphImageFormat, InvalidFont, Outline, Rect};
+use crate::{point, v2, Font, GlyphId, GlyphImageFormat, GlyphSvg, InvalidFont, Outline, Rect};
 use alloc::boxed::Box;
 #[cfg(not(feature = "std"))]
 use alloc::vec::Vec;
@@ -172,6 +172,7 @@ impl FontVec {
     /// assert_eq!(font.into_vec(), font_data_clone);
     /// # Ok(()) }
     /// ```
+    #[inline]
     pub fn into_vec(self) -> Vec<u8> {
         self.0.face.into_vec()
     }
@@ -340,6 +341,21 @@ macro_rules! impl_font {
                         ttfp::RasterImageFormat::BitmapPremulBgra32 => BitmapPremulBgra32,
                     },
                 })
+            }
+
+            fn glyph_svg_image(&self, id: GlyphId) -> Option<GlyphSvg> {
+                let img = self.0.as_face_ref().glyph_svg_image(id.into())?;
+
+                Some(GlyphSvg {
+                    data: img.data,
+                    start_glyph_id: GlyphId(img.start_glyph_id.0),
+                    end_glyph_id: GlyphId(img.end_glyph_id.0),
+                })
+            }
+
+            #[inline]
+            fn font_data(&self) -> &[u8] {
+                self.0.as_face_ref().raw_face().data
             }
         }
     };